Refactor MarkdownV2 styling helpers to separate files for bold, italic, strikethrough, blockquote, and code block styles
