+++ benchmarks/StringEnricher.Benchmarks/Nodes/SingleBoldNodeBenchmarks/BoldNodeBenchmarks.cs
@@ -2,6 +2,7 @@
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Order;
 using StringEnricher.Benchmarks.AnalogueImplementations;
+using StringEnricher.Helpers.MarkdownV2;
 using StringEnricher.Nodes.MarkdownV2.Formatting;
 
 namespace StringEnricher.Benchmarks.Nodes.SingleBoldNodeBenchmarks;
@@ -24,28 +25,52 @@ public class BoldNodeBenchmarks
     }
 
     [Benchmark]
-    public string BoldMarkdownV2_Apply() => BoldMarkdownV2.Apply(textToBold).ToString();
+    public string BoldMarkdownV2_Apply()
+    {
+        return BoldMarkdownV2.Apply(textToBold).ToString();
+    }
 
     [Benchmark]
-    public string MessageTextStyleLambda_Bold() => MessageTextNodeLambda.Bold.Build(textToBold);
+    public string MessageTextStyleLambda_Bold()
+    {
+        return MessageTextNodeLambda.Bold.Build(textToBold);
+    }
 
     [Benchmark]
-    public string MessageTextStyleStringHandler_Bold() => MessageTextNodeStringHandler.Bold.Build(textToBold);
+    public string MessageTextStyleStringHandler_Bold()
+    {
+        return MessageTextNodeStringHandler.Bold.Build(textToBold);
+    }
 
     [Benchmark(Baseline = true)]
-    public string InterpolatedString() => $"*{textToBold}*";
+    public string InterpolatedString()
+    {
+        return $"*{textToBold}*";
+    }
 
     [Benchmark]
-    public string StringFormat() => string.Format("*{0}*", textToBold);
+    public string StringFormat()
+    {
+        return string.Format("*{0}*", textToBold);
+    }
 
     [Benchmark]
-    public string Concatenation() => "*" + textToBold + "*";
+    public string Concatenation()
+    {
+        return "*" + textToBold + "*";
+    }
 
     [Benchmark]
-    public string StringJoin() => string.Join("", "*", textToBold, "*");
+    public string StringJoin()
+    {
+        return string.Join("", "*", textToBold, "*");
+    }
 
     [Benchmark]
-    public string StringConcat() => string.Concat("*", textToBold, "*");
+    public string StringConcat()
+    {
+        return string.Concat("*", textToBold, "*");
+    }
 
     [Benchmark]
     public string StringBuilder_Default()
