+++ benchmarks/StringEnricher.Benchmarks/Nodes/TenfoldBoldNodeBenchmarks/TenfoldBoldNodeBenchmarks.cs
@@ -2,6 +2,7 @@ using System.Text;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Order;
 using StringEnricher.Benchmarks.AnalogueImplementations;
+using StringEnricher.Helpers.MarkdownV2;
 using StringEnricher.Nodes.MarkdownV2.Formatting;
 
 namespace StringEnricher.Benchmarks.Nodes.TenfoldBoldNodeBenchmarks;
