+++ tests/StringEnricher.Tests/CompositeNodeTests.cs
@@ -1,4 +1,5 @@
-﻿using StringEnricher.Nodes.MarkdownV2.Formatting;
+﻿using StringEnricher.Helpers.MarkdownV2;
+using StringEnricher.Nodes.MarkdownV2.Formatting;
 using StringEnricher.Nodes.Shared;
 
 namespace StringEnricher.Tests;
