+++ tests/StringEnricher.Tests/MarkdownV2/Nodes/BoldNodeTests.cs
@@ -1,3 +1,4 @@
+using StringEnricher.Helpers.MarkdownV2;
 using StringEnricher.Nodes.MarkdownV2.Formatting;
 
 namespace StringEnricher.Tests.MarkdownV2.Nodes;
