+++ tests/StringEnricher.Tests/MarkdownV2/Nodes/ExpandableBlockquoteNodeTests.cs
@@ -1,3 +1,4 @@
+using StringEnricher.Helpers.MarkdownV2;
 using StringEnricher.Nodes.MarkdownV2.Formatting;
 using StringEnricher.Nodes.Shared;
