+++ tests/StringEnricher.Tests/MarkdownV2/Nodes/InlineLinkNodeTests.cs
@@ -1,4 +1,5 @@
-﻿using StringEnricher.Nodes.MarkdownV2.Formatting;
+﻿using StringEnricher.Helpers.MarkdownV2;
+using StringEnricher.Nodes.MarkdownV2.Formatting;
 
 namespace StringEnricher.Tests.MarkdownV2.Nodes;
